CasiAcabo

+++ src/java/VO/cursopersonaVO.java
@@ -2,22 +2,22 @@ package VO;
 
 public class cursopersonaVO {
  
-	private int id_persona;
-	private int id_curso;
+	private long id_persona;
+	private long id_curso;
 
-	public int getid_persona() {
+	public long getid_persona() {
 		return this.id_persona;
 	}
 
-	public void setid_persona(int id_persona) {
+	public void setid_persona(long id_persona) {
 		this.id_persona=id_persona;
 	}
 
-	public int getid_curso() {
+	public long getid_curso() {
 		return this.id_curso;
 	}
 
-	public void setid_curso(int id_curso) {
+	public void setid_curso(long id_curso) {
 		this.id_curso=id_curso;
 	}
